requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apideck-unify==0.14.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
cli_exit_tools==1.2.7
click==8.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpath-python==1.0.6
lib-detect-testenv==2.0.8
lib-shopware6-api==2.0.6
lib_shopware6_api_base==2.1.9
MarkupSafe==3.0.2
oauthlib==3.2.2
pydantic==2.11.4
pydantic_core==2.33.2
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
